feat(form-restoration): Form restoration now works
When you navigate away from a page and then return to it there is a
"form restoration" process done by the browser. It puts values back as
they were, by filling in the <select>'s and <input>'s (though not
<input type="hidden">) with their old values.

By using this fact I've created a mechanism where the map saves it's
lat, lng and zoom levels in a invisible <input>, and using the pageshow
event to utilize this data after a navigation.

There is also work on problem that the center of the map has been used
instead of marker position.